Add manufacturers paging

+++ app/src/main/java/my/illrock/a1codechallenge/data/network/response/manufacturer/ManufacturersResponse.kt
@@ -7,5 +7,5 @@ data class ManufacturersResponse(
     val page: Int,
     val pageSize: Int,
     val totalPageCount: Int,
-    val wkda: Map<Int, String>
+    val wkda: Map<Long, String>
 )
